requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
openai==1.55.3
pydantic==2.5.0
duckduckgo-search==0.9
python-dotenv==1.0.0
requests==2.31.0
langchain==0.0.340
langchain-openai==0.0.2
azure-identity==1.15.0
httpx==0.27.2
pytest==7.4.4